+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## 🛑 Breaking changes 🛑
 
+- Use default/hardcoded configuration #1273.
 - Remove `reconnection_delay` from OpenCensus exporter #1516.
 - Remove old receiver factories and receiver base factory #1583.
 - Remove logs factories and merge with normal factories #1569.